requirements.txt
networkx
pyyaml
jinja2